backend/requirements.txt
uvicorn==0.19.0
gunicorn==20.1.0
fastapi==0.85.1
mysql-connector-python==8.0.31
pylint==2.15.5
coverage==6.5.0
requests==2.28.1
python-jose[cryptography]==3.3.0
validate_email==1.3
passlib~=1.7.4
python-multipart~=0.0.5
py3dns~=3.2.1
algoliasearch~=2.6.3